Changed delta time calculation back to double to increase accuracy up until the last moment.

+++ engine/platform/glfw/platform_glfw_interface.h
@@ -50,7 +50,7 @@ private:
 
     bool initialized_;
     bool running_;
-    float previous_time_;
+    double previous_time_;
 };
 
 } // namespace glfw
